requirements.txt
PyMuPDF==1.24.5
Pillow==10.3.0
pandas==2.2.2

langchain==0.2.5
langchain_openai==0.1.9
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
pydantic-settings==2.3.3
python-multipart==0.0.9
itsdangerous==2.2.0
apscheduler==3.10.4
colorlog==6.8.2

oauth2client==4.1.3
gspread==6.1.2
gspread_dataframe==4.0.0
google-auth==2.30.0
google-api-python-client==2.104.0
